requirements.txt
pdfCropMargins==1.0.9
pydantic==1.9.2
pyqt5==5.15.7
win32printing==0.1.3
zebra==0.1.0
pywin32==304